refactor(matrix): refactor matrix event checking
- AmuCtrlRecorder/AmuCtrlChecker
- AmuExecRecorder/AmuExecChecker
- TokenRecorder/TokenChecker
