+++ src/test/csrc/difftest/diffstate.h
@@ -19,6 +19,7 @@
 
 #include "common.h"
 #include <cstdint>
+#include <deque>
 #include <queue>
 #include <unordered_set>
 
@@ -148,6 +149,49 @@ public:
   std::queue<DifftestLoadEvent> load_event_queue;
 #endif // CONFIG_DIFFTEST_LOADEVENT
 #endif // CONFIG_DIFFTEST_SQUASH
+
+#ifdef CONFIG_DIFFTEST_AMUCTRLEVENT
+  // AME instruction lifecycle in software ROB
+  // - An AME instruction is first committed in the DUT's hardware ROB. At that
+  // moment, it is captured by AmuCtrlRecorder and from then on it is no
+  // longer managed by the DUT hardware ROB, but instead by the DiffTest
+  // framework's software ROB. In this phase, the instruction waits for the REF
+  // side to commit it as well; this corresponds to the WAIT_REF_COMMIT state.
+  // - Once REF has also committed this instruction, AmuCtrlChecker compares
+  // the amu_ctrl signals from DUT and REF. If the signals match, the state
+  // transitions to WAIT_DUT_EXEC, meaning we are now waiting for the DUT's
+  // execution result.
+  // - Next, the DUT's matrix unit executes the instruction and gradually
+  // writes the result into the matrix registers. amu_inst_finish_record
+  // captures this writeback process and updates the mirrored registers in
+  // DiffTest. When the instruction has fully finished and all matrix register
+  // writes are complete, the state becomes WAIT_SWROB_COMMIT, waiting for the
+  // software ROB to commit the instruction.
+  // - An instruction in the software ROB can be committed if and only if it
+  // has completed and all preceding instructions have also completed.
+  // do_amuexec_check commits such ready instructions: the REF side re-executes
+  // the instruction and compares its result against the DUT result. If they
+  // match, the lifecycle of this instruction ends and its resources can be
+  // released.
+  enum AmeInstState {
+    INVALID = 0,
+    WAIT_REF_COMMIT,
+    WAIT_DUT_EXEC,
+    WAIT_SWROB_COMMIT
+  };
+
+  typedef struct {
+    DifftestAmuCtrlEvent amu_event;
+    AmeInstState state;
+    uint64_t *res;
+  } AmeInstRobEntry;
+
+  std::deque<AmeInstRobEntry> matrix_sw_rob;
+#endif // CONFIG_DIFFTEST_AMUCTRLEVENT
+
+#ifdef CONFIG_DIFFTEST_TOKENEVENT
+  std::queue<DifftestTokenEvent> token_event_queue;
+#endif // CONFIG_DIFFTEST_TOKENEVENT
 
 #ifdef DEBUG_REFILL
   uint64_t track_instr = 0;
