+++ src/test/csrc/difftest/difftest.cpp
@@ -242,7 +242,6 @@ Difftest::Difftest(int coreid) {
 #endif // CONFIG_DIFFTEST_REPLAY
 
 #ifdef CONFIG_DIFFTEST_AMUCTRLEVENT
-  have_check_ctrl = matrix_sw_rob.begin();
   mma_verifier = nullptr;
   
   // Initialize AMU finish event buffers
@@ -408,6 +407,27 @@ void Difftest::init_checkers() {
       [this]() -> DifftestSyncCustomMflushpwrEvent & { return dut->sync_custom_mflushpwr; }, state, proxy));
 #endif
 
+#ifdef CONFIG_DIFFTEST_AMUCTRLEVENT
+  for (int i = 0; i < CONFIG_DIFF_AMU_CTRL_WIDTH; i++) {
+    checkers.push_back(new AmuCtrlRecorder(
+        [this, i]() -> DifftestAmuCtrlEvent & { return dut->amu_ctrl[i]; }, state, proxy));
+  }
+  checkers.push_back(new AmuCtrlChecker(state, proxy));
+  for (int i = 0; i < CONFIG_DIFF_AMU_FINISH_WIDTH; i++) {
+    checkers.push_back(new AmuExecRecorder(
+        [this, i]() -> DifftestAmuFinishEvent & { return dut->amu_finish[i]; }, state, proxy));
+  }
+  checkers.push_back(new AmuExecChecker(state, proxy));
+#endif // CONFIG_DIFFTEST_AMUCTRLEVENT
+
+#ifdef CONFIG_DIFFTEST_TOKENEVENT
+  for (int i = 0; i < CONFIG_DIFF_TOKEN_WIDTH; i++) {
+    checkers.push_back(new TokenRecorder(
+        [this, i]() -> DifftestTokenEvent & { return dut->token[i]; }, state, proxy));
+  }
+  checkers.push_back(new TokenChecker(state, proxy));
+#endif // CONFIG_DIFFTEST_TOKENEVENT
+
   arch_event_checker = new ArchEventChecker([this]() -> DifftestArchEvent & { return dut->event; }, state, proxy,
                                             [this]() -> const DiffTestRegState & { return dut->regs; });
 
@@ -497,9 +517,12 @@ void Difftest::do_replay() {
     state->load_event_queue.pop();
 #endif
 #ifdef CONFIG_DIFFTEST_AMUCTRLEVENT
-  while (!matrix_sw_rob.empty())
-    matrix_sw_rob.pop();
+  state->matrix_sw_rob.clear();
 #endif // CONFIG_DIFFTEST_AMUCTRLEVENT
+#ifdef CONFIG_DIFFTEST_TOKENEVENT
+  while (!state->token_event_queue.empty())
+    state->token_event_queue.pop();
+#endif // CONFIG_DIFFTEST_TOKENEVENT
 }
 #endif // CONFIG_DIFFTEST_REPLAY
 
@@ -581,12 +604,6 @@ inline int Difftest::check_all() {
     }
   }
 
-#ifdef CONFIG_DIFFTEST_AMUCTRLEVENT
-  amu_ctrl_event_record();
-  amu_inst_finish_record();
-  token_event_record();
-#endif // CONFIG_DIFFTEST_AMUCTRLEVENT
-
 #ifdef DEBUG_MODE_DIFF
   // skip load & store insts in debug mode
   // for other insts copy inst content to ref's dummy debug module
@@ -622,18 +639,6 @@ inline int Difftest::check_all() {
     }
   }
 
-#ifdef CONFIG_DIFFTEST_AMUCTRLEVENT
-  if (do_amuctrl_check()) {
-    return 1;
-  }
-  if (do_amuexec_check()) {
-    return 1;
-  }
-  if (do_token_check()) {
-    return 1;
-  }
-#endif // CONFIG_DIFFTEST_AMUCTRLEVENT
-
   if (int ret = update_delayed_writeback()) {
     return ret;
   }
@@ -668,306 +673,6 @@ inline int Difftest::check_all() {
 
   return DiffTestChecker::STATE_OK;
 }
-
-#ifdef CONFIG_DIFFTEST_AMUCTRLEVENT
-static char amu_ctrl_op_str[4][16] = {"MMA", "MLS", "MRELEASE", "MARITH"};
-#endif // CONFIG_DIFFTEST_AMUCTRLEVENT
-
-#ifdef CONFIG_DIFFTEST_AMUCTRLEVENT
-int Difftest::do_amuctrl_check() {
-  for (auto iter = matrix_sw_rob.begin(); iter != matrix_sw_rob.end(); ++iter) {
-    if (iter->state == WAIT_REF_COMMIT) {
-      DifftestAmuCtrlEvent amu_event = iter->amu_event;
-  #ifdef CONFIG_DIFFTEST_SQUASH
-      // TODO: What is squash? How to squash?
-  #endif // CONFIG_DIFFTEST_SQUASH
-      // Save the amu event info
-      auto rm = amu_event.rm;
-      auto op = amu_event.op;
-      auto md = amu_event.md;
-      auto sat = amu_event.sat;
-      auto ms1 = amu_event.ms1;
-      auto ms2 = amu_event.ms2;
-      auto mtilem = amu_event.mtilem;
-      auto mtilen = amu_event.mtilen;
-      auto mtilek = amu_event.mtilek;
-      auto types1 = amu_event.types1;
-      auto types2 = amu_event.types2;
-      auto typed = amu_event.typed;
-      auto transpose = amu_event.isfp;
-      auto base = amu_event.base;
-      auto stride = amu_event.stride;
-      auto isfp = amu_event.isfp;
-      uint64_t pc = amu_event.pc;
-
-      int check_res = proxy->get_amu_ctrl_event(&amu_event);
-
-      if (check_res == 1) {
-        // Compare the amu event info
-        // For dismatch, proxy returns 1 and sets the amu event info
-        display();
-
-        printf("\n==============  Amu Mma Ctrl Event (Core %d)  ==============\n", state->coreid);
-        printf("Mismatch for amu ctrl event \n");
-        printf("  REF AMU ctrl: pc 0x%016lx, op %s\n", amu_event.pc, amu_ctrl_op_str[amu_event.op]);
-        switch (amu_event.op) {
-          case 0: // MMA
-            printf("                md %d, rm %d, sat %d, ms1 %d, ms2 %d, fp %d,\n"
-                  "                mtilem %d, mtilen %d, mtilek %d, types1 %d, types2 %d, typed %d\n",
-                  amu_event.md, amu_event.rm, amu_event.sat, amu_event.ms1, amu_event.ms2, amu_event.isfp,
-                  amu_event.mtilem, amu_event.mtilen, amu_event.mtilek, amu_event.types1, amu_event.types2, amu_event.typed);
-            break;
-          case 1: // MLS
-            printf("                md %d, ls %d, transpose %d, isacc %d,\n"
-                  "                base %016lx, stride %016lx, row %d, column %d, widths %d\n",
-                  amu_event.md, amu_event.sat, amu_event.isfp, amu_event.types1,
-                  amu_event.base, amu_event.stride, amu_event.mtilem, amu_event.mtilen, amu_event.typed);
-            break;
-          case 2: // MRelease
-            printf("                tokenRd %d\n", amu_event.mtilem);
-            break;
-          case 3: // Arith
-            printf("                md %d, opType %#lx\n", amu_event.md, amu_event.base);
-            break;
-          default:
-            printf("                Unknown amu event op\n");
-            break;
-        }
-        printf("  DUT AMU ctrl: pc 0x%016lx, op %s\n", pc, amu_ctrl_op_str[op]);
-        switch (op) {
-          case 0: // MMA
-            printf("                md %d, rm %d, sat %d, ms1 %d, ms2 %d, fp %d,\n"
-                  "                mtilem %d, mtilen %d, mtilek %d, types1 %d, types2 %d, typed %d\n",
-                  md, rm, sat, ms1, ms2, isfp, mtilem, mtilen, mtilek, types1, types2, typed);
-            break;
-          case 1: // MLS
-            printf("                md %d, ls %d, transpose %d, isacc %d,\n"
-                  "                base %016lx, stride %016lx, row %d, column %d, widths %d\n",
-                  md, sat, transpose, types1, base, stride, mtilem, mtilen, typed);
-            break;
-          case 2: // MRelease
-            printf("                tokenRd %d\n", mtilem);
-            break;
-          case 3: // Arith
-            printf("                md %d, opType %#lx\n", md, base);
-            break;
-          default:
-            printf("                Unknown amu event op\n");
-            break;
-        }
-        return 1;
-      } else if (check_res == -1) {
-        display();
-        printf("\n==============  Amu Mma Ctrl Event (Core %d)  ==============\n", state->coreid);
-        printf("  No available REF AMU ctrl\n");
-        printf("  DUT AMU ctrl: pc 0x%016lx, op %s\n", pc, amu_ctrl_op_str[op]);
-        switch (op) {
-          case 0: // MMA
-            printf("                md %d, rm %d, sat %d, ms1 %d, ms2 %d, fp %d,\n"
-                  "                mtilem %d, mtilen %d, mtilek %d, types1 %d, types2 %d, typed %d\n",
-                  md, rm, sat, ms1, ms2, isfp, mtilem, mtilen, mtilek, types1, types2, typed);
-            break;
-          case 1: // MLS
-            printf("                md %d, ls %d, transpose %d, isacc %d,\n"
-                  "                base %016lx, stride %016lx, row %d, column %d, widths %d\n",
-                  md, sat, transpose, types1, base, stride, mtilem, mtilen, typed);
-            break;
-          case 2: // MRelease
-            printf("                tokenRd %d\n", mtilem);
-            break;
-          case 3: // Arith
-            printf("                md %d, opType %#lx\n", md, base);
-            break;
-          default:
-            printf("                Unknown amu event op\n");
-            break;
-        }
-        return 1;
-      } else {
-        iter->state = WAIT_DUT_EXEC;
-      }
-    }
-  }
-  return 0;
-}
-#endif // CONFIG_DIFFTEST_AMUCTRLEVENT
-
-#ifdef CONFIG_DIFFTEST_AMUCTRLEVENT
-int Difftest::do_amuexec_check() {
-  // For each amu ctrl in sw_rob, check whether the inst is able to be committed. 
-  for (auto iter = matrix_sw_rob.begin(); iter != matrix_sw_rob.end(); ) {
-    if (iter->state == WAIT_SWROB_COMMIT) {
-      DifftestAmuCtrlEvent amu_event = iter->amu_event;
-      uint8_t op = amu_event.op;
-      MmaVerificationBuffer *buffer = nullptr;
-      switch (op) {
-        case 0: // MMA
-          // Allocate buffer for MMA verification
-          buffer = mma_verifier->allocate_buffer(&amu_event);
-          // Store DUT result in the buffer
-          memcpy(buffer->dut_result, iter->res, amu_event.mtilem * amu_event.mtilen * get_element_size(amu_event.typed));
-          // Call get_amu_lazy with buffer pointers
-          proxy->get_amu_lazy(&amu_event, iter->res, buffer->src1, buffer->src2, buffer->src3);
-          // Pass buffer to verification thread
-          mma_verifier->add_to_verification_queue(buffer);
-          break;
-        case 1: // MLS
-        case 2: // MRelease
-        case 3: // Arith
-          if (proxy->get_amu_exec(&amu_event, iter->res) == 1) {
-            // Dismatch
-            display();
-            printf("Mismatch for amu exec event: pc 0x%016lx, op %s\n", amu_event.pc, amu_ctrl_op_str[amu_event.op]);
-            return 1;
-          }
-          break;
-        default:
-          printf("Unknown amu event op: %d\n", op);
-          return 1;
-          break;
-      }
-      // Remove the processed event from matrix_sw_rob
-      iter = matrix_sw_rob.erase(iter);
-    } else {
-      // When there's an unfinished inst before the current inst, the current inst will not be committed.
-      // So, when the current inst is unfinished, break the loop.
-      break;
-    }
-  }
-  return 0;
-}
-#endif // CONFIG_DIFFTEST_AMUCTRLEVENT
-
-#ifdef CONFIG_DIFFTEST_TOKENEVENT
-static char token_event_op_str[2][16] = {"msyncregreset", "macquire"};
-#endif // CONFIG_DIFFTEST_TOKENEVENT
-
-#ifdef CONFIG_DIFFTEST_TOKENEVENT
-int Difftest::do_token_check() {
-  while (!token_event_queue.empty()) {
-    DifftestTokenEvent token_event = token_event_queue.front();
-#ifdef CONFIG_DIFFTEST_SQUASH
-    // TODO: What is squash? How to squash?
-#endif // CONFIG_DIFFTEST_SQUASH
-    // Save the token event info
-    auto op = token_event.op;
-    auto tokenRd = token_event.tokenRd;
-    uint64_t pc = token_event.pc;
-
-    int check_res = proxy->get_token_event(&token_event);
-
-    if (check_res == 1) {
-      // Compare the token event info
-      // For dismatch, proxy returns 1 and sets the token event info
-      display();
-      printf("\n==============  Token Event (Core %d)  ==============\n", state->coreid);
-      printf("Mismatch for token event\n");
-      printf("  REF Token: pc 0x%016lx, op %s, tokenRd %d\n", token_event.pc, token_event_op_str[token_event.op], token_event.tokenRd);
-      printf("  DUT Token: pc 0x%016lx, op %s, tokenRd %d\n", pc, token_event_op_str[op], tokenRd);
-      token_event_queue.pop();
-      return 1;
-    } else if (check_res == -1) {
-      display();
-      printf("\n==============  Token Event (Core %d)  ==============\n", state->coreid);
-      printf("  No available REF Token\n");
-      printf("  DUT Token: pc 0x%016lx, op %s, tokenRd %d\n", pc, token_event_op_str[op], tokenRd);
-      token_event_queue.pop();
-      return 1;
-    }
-    token_event_queue.pop();
-  }
-  return 0;
-}
-#endif // CONFIG_DIFFTEST_TOKENEVENT
-
-#ifdef CONFIG_DIFFTEST_AMUCTRLEVENT
-void Difftest::amu_ctrl_event_record() {
-  for (int i = 0; i < CONFIG_DIFF_AMU_CTRL_WIDTH; i++) {
-    if (dut->amu_ctrl[i].valid) {
-      AmeInstRobEntry entry;
-      entry.amu_event = dut->amu_ctrl[i];
-      entry.state = WAIT_REF_COMMIT;
-      entry.res = NULL;
-      matrix_sw_rob.push_back(entry);
-      dut->amu_ctrl[i].valid = 0;
-    }
-  }
-}
-
-void Difftest::amu_inst_finish_record() {
-  const static size_t ARLen = 128 * 32;
-  const static size_t TRLen = 64 * 8;
-  const static size_t ABankWidth = 256;
-  const static size_t TBankWidth = 256;
-  for (int i = 0; i < CONFIG_DIFF_AMU_FINISH_WIDTH; ++i) {
-    if (dut->amu_finish[i].valid) {
-      for (auto &entry : matrix_sw_rob) {
-        if (entry.amu_event.pc == dut->amu_finish[i].pc && entry.state == WAIT_DUT_EXEC) {
-          if (entry.res == NULL) {
-            // first `valid` for this inst
-            // alloc space for matrix inst
-            size_t matrix_size = 0;
-            switch (entry.amu_event.op) {
-              // TODO: do not use hardcoded matrix size here
-              case 0: // MMA
-                matrix_size = 128 * 128 * 4;
-                break;
-              case 1: // Matrix load/store
-              case 3: // Matrix Arith
-                if (entry.amu_event.sat == 0) { // load
-                  if ((entry.amu_event.md & 4) == 0) { // tile register
-                    matrix_size = 128 * 64 * 1;
-                  } else { // acc register
-                    matrix_size = 128 * 128 * 4;
-                  }
-                }
-                break;
-              default:
-                break;
-            }
-            entry.res = new uint64_t[matrix_size / 8];
-            memset(entry.res, 0, matrix_size);
-          }
-          uint8_t md = entry.amu_event.md;
-          size_t stride = 0;
-          if (md < 4) {
-            stride = TRLen / TBankWidth;
-          } else {
-            stride = ARLen / ABankWidth;
-          }
-          for (int j = 0; j < 8; ++j) { //  for each bank
-            if (dut->amu_finish[i].bankValid[j]) {
-              size_t addr = dut->amu_finish[i].bankAddr[j];
-              for (int k = 0; k < 4; ++k) {
-                entry.res[(addr / stride * stride * 8 + j * stride + addr % stride) * 4 + k] =
-                  dut->amu_finish[i].data[j * 4 + k];
-              }
-            }
-          }
-          
-          if (dut->amu_finish[i].finish) {
-            // last `valid` for this inst
-            entry.state = WAIT_SWROB_COMMIT;
-          }
-          break;
-        }
-      }
-      dut->amu_finish[i].valid = 0;
-    }
-  }
-}
-#endif // CONFIG_DIFFTEST_AMUCTRLEVENT
-
-#ifdef CONFIG_DIFFTEST_TOKENEVENT
-void Difftest::token_event_record() {
-  for (int i = 0; i < CONFIG_DIFF_TOKEN_WIDTH; i++) {
-    if (dut->token[i].valid) {
-      token_event_queue.push(dut->token[i]);
-      dut->token[i].valid = 0;
-    }
-  }
-}
-#endif // CONFIG_DIFFTEST_TOKENEVENT
 
 int Difftest::update_delayed_writeback() {
 #define CHECK_DELAYED_WB(wb, delayed, n, regs_name)                                                \
